feat(start): Initial commit
This is my initial commit of a Vagrantfile. It doesn't do a ton
right now, ports are just 80 and 8080. Software installed is Node v4
npm, gulp, babel, browserify, and go.

I might want to think a little bit more about the use of go in here as
it just turns it into a general virtual machine, but I'm fine with that for
the time being.

Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>